Backend Complete

+++ backend/controller/product/controller.js
@@ -55,11 +55,14 @@ export const DeleteProduct = async (req, res) => {
   const { id } = req.params;
   // console.log("Id : ", id);
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ success: false, message: "Invalid Product ID" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product Deleted" });
   } catch (error) {
     console.log("Error in deleting Products", error.message);
-    res.status(400).json({ success: false, message: "Product Not Found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
